TeamCode + Members
Get Team ID
Members
Disable field as mem
