+++ web/package.json
@@ -24,6 +24,7 @@
     "moment": "^2.29.1",
     "query-string": "^6.14.1",
     "react": "^17.0.1",
+    "react-copy-to-clipboard": "^5.0.3",
     "react-dom": "^17.0.1",
     "react-horizontal-timeline": "^1.5.3",
     "react-icons": "^4.2.0",
@@ -33,6 +34,7 @@
     "react-scripts": "4.0.2",
     "socket.io-client": "^3.1.2",
     "socketio-jwt": "^4.6.2",
+    "styled-components": "^5.3.0",
     "web-vitals": "^1.1.0"
   },
   "scripts": {
