+++ web/src/components/Team/Members.js
@@ -24,25 +24,9 @@ const API_URL = config.API_LOCAL;
 const MembersOfTeam = (props) => {
   const classes = useStyles();
 
-  const teamID = props.teamID;
-  const { isLoggedIn } = useContext(MyContext);
-  const [members, setMembers] = useState([]);
-  const history = useHistory();
+  const {members, isAdmin, teamID, getTeamMembers} = props;
   const token = localStorage.getItem('jwtToken');
   const userID = localStorage.getItem('userID');
-  const [isAdmin, setIsAdmin] = useState(false);
-
-  useEffect(() => {
-    console.log(isLoggedIn);
-    if (isLoggedIn !== null && isLoggedIn === false) {
-      history.push('/');
-    }
-    getTeamMembers();
-  }, [isLoggedIn]);
-
-  useEffect(() => {
-    isAdminTeam();
-  }, [members]);
 
   const handleRemoveMember = async (userID) => {
     console.log("remove" + userID);
@@ -56,55 +40,7 @@ const MembersOfTeam = (props) => {
         Authorization: `Bearer ${token}`,
       },
       body: JSON.stringify(data)
-    });
-
-  }
-
-  const getTeamMembers = async () => {
-
-    const res = await fetch(`${API_URL}/teams/${teamID}/members`, {
-      method: 'GET',
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${token}`,
-      }
-    });
-    console.log(res.body);
-    if (res.status === 200) {
-      const result = await res.json();
-      console.log(result.members);
-      setMembers(result.members);
-    } else {
-      // alert("Some error when updating!")
-    }
-  }
-
-  const isAdminTeam = () => {
-    console.log("userID" + userID);
-    console.log('mem', members);
-    // const fakeArray = members.slice();
-    // fakeArray
-    //   .filter(member => member.ID === userID)
-    //   .map(member => {
-    //     console.log("Role: " + member.Role)
-    //     if (member.Role === 1)
-    //       setIsAdmin(true);
-    //     else setIsAdmin(false)
-    //   })
-
-    let _isAdmin = false;
-
-    members.forEach((mem) => {
-
-      if (mem.ID === userID) {
-        if (mem.Role === 1) {
-          _isAdmin = true;
-          return;
-        }
-      }
-    });
-    console.log(_isAdmin);
-    setIsAdmin(_isAdmin);
+    }).then(getTeamMembers);
   }
 
   return (
